refact: move DocumentInputStream to internal package
